Separate platform module from utils
